encapsulate the `Region` struct within region inference
